+++ src/com/company/repo/repoImpls.java
@@ -1,0 +1,4 @@
+package com.company.repo;
+
+public class repoImpls {
+}
